Factor out email sending into its own service
- for later reuse.
